fix(v3): extract the negotiation fallback schema to a named class
phpstan does not read the inline @var/@param PHPDoc on the anonymous
TabularSchema the negotiation middleware built for its raw-array fallback,
so a fresh analysis (the 3.x into master diff) flagged the $columns
property, the constructor parameter, and the columns() return as untyped
arrays. Named classes resolve their member PHPDoc reliably, so the schema
is now a small @internal DecodedPayloadSchema and the types are honoured.
